+++ spec/app/models/image.rb
@@ -1,15 +1,20 @@
 class Image
   include Mongoid::Fields::Serializable
 
+  attr_reader :name
+
   def initialize(name)
     @name = name
   end
 
   def deserialize(value)
-    value.to_s
+    self.class.new(value)
   end
 
   def serialize(value)
-    value.to_s
+    value.name.to_s
   end
+end
+
+class Thumbnail < Image
 end