+++ spec/app/models/movie.rb
@@ -2,6 +2,7 @@ class Movie
   include Mongoid::Document
   field :title, :type => String
   field :poster, :type => Image
+  field :poster_thumb, :type => Thumbnail
   has_many :ratings, :as => :ratable, :dependent => :nullify
   has_many :comments